Mostrar en cuanto aparece un alumno

+++ android/TurnoClase/turnoclaseprofesor/src/main/java/com/jaureguialzo/turnoclaseprofesor/MainActivity.kt
@@ -206,6 +206,7 @@ class MainActivity : AppCompatActivity() {
                                 Log.e(TAG, "Error al recuperar datos: ", error)
                             } else {
                                 actualizarAula(querySnapshot!!.documents.count())
+                                mostrarSiguiente()
                             }
                         }
                     }
@@ -218,7 +219,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private fun botonSiguiente() {
+    private fun mostrarSiguiente(avanzarCola: Boolean = false) {
 
         Log.d(TAG, "Mostrando el siguiente alumno...")
 
@@ -247,7 +248,9 @@ class MainActivity : AppCompatActivity() {
                                                         actualizarMensaje(alumno["nombre"] as String)
 
                                                         // Borrar la entrada de la cola
-                                                        refPosicion.delete()
+                                                        if (avanzarCola) {
+                                                            refPosicion.delete()
+                                                        }
                                                     } else {
                                                         Log.e(TAG, "El alumno no existe")
                                                         actualizarMensaje("?")
@@ -266,12 +269,15 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
+    private fun botonSiguiente() {
+        mostrarSiguiente(true)
+    }
+
     private fun botonEnCola() {
 
-        Log.d("TurnoClase", "Este botón ya no hace nada :)")
+        Log.d("TurnoClase", "Este botón sólo se usa para los test de UI")
 
         if (isRunningTest) {
-            n -= 1
             if (n >= 0) {
                 actualizarAula(n)
                 actualizarMensaje(Nombres().aleatorio())
@@ -279,15 +285,19 @@ class MainActivity : AppCompatActivity() {
                 actualizarAula(0)
                 actualizarMensaje("")
             }
+
+            n -= 1
         }
     }
 
     private fun botonCodigoAulaCorto() {
 
-        Log.d(TAG, "Vaciando el aula...")
-
+        //Log.d(TAG, "Vaciando el aula...")
         // Pendiente de implementar en el servidor, no se puede borrar una colección desde el cliente
         // REF: https://firebase.google.com/docs/firestore/manage-data/delete-data?hl=es-419
+
+        // Menú de acciones para gestionar múltiples profesores
+        //mostrarAcciones()
 
     }
 
@@ -318,13 +328,13 @@ class MainActivity : AppCompatActivity() {
 
     }
 
-    private fun actualizarAula(codigoAula: String = "?", enCola: Int = -1) {
-        actualizarAula(codigoAula)
+    private fun actualizarAula(codigo: String = "?", enCola: Int = -1) {
+        actualizarAula(codigo)
         actualizarAula(enCola)
     }
 
-    private fun actualizarAula(codigoAula: String) {
-        botonCodigoAula.text = codigoAula
+    private fun actualizarAula(codigo: String) {
+        botonCodigoAula.text = codigo
     }
 
     private fun actualizarAula(enCola: Int) {
